misc: Now also send info about the fake player to the disguising player
