+++ src/main/java/xyz/nifeather/morph/backends/server/renderer/network/datawatcher/watchers/SingleWatcher.java
@@ -651,7 +651,7 @@ public abstract class SingleWatcher extends MorphPluginObject
 
     public abstract boolean haveAnimation(WrapperPlayServerEntityAnimation.EntityAnimationType animationType);
 
-    public void beginSelfUpdateIfAny()
+    public void onDisguiseApply()
     {
     }
