+++ src/main/java/xyz/nifeather/morph/backends/server/renderer/network/datawatcher/watchers/types/RecoveringPlayerWatcher.java
@@ -1,0 +1,11 @@
+package xyz.nifeather.morph.backends.server.renderer.network.datawatcher.watchers.types;
+
+import org.bukkit.entity.Player;
+
+public class RecoveringPlayerWatcher extends AbstractPlayerWatcher
+{
+    public RecoveringPlayerWatcher(Player bindingPlayer)
+    {
+        super(bindingPlayer);
+    }
+}
